Remove syntax provider and refactor useLokiLabels
